Expect user to close connection abruptly
When waiting for input from socket, expect not only a valid outcome,
but also the client getting bored and closing the connection without
any message being sent.Log an info message when that happens,but do
not make it an error.

Also, only use hyphens for headers sent to registry server.